Initial upload Timex Data Link Docking station
Files for the 3D print of the Timex Data Link 150S Docking Station, using a Pi Pico to simulate the Timex Data Link Notebook Adapter.
The docking station requires two foam pads fixed to the top and bottom of the watch aperture (to prevent scratching). The pads I used are cut from 3.5mm foam (available on Amazon in a roll). The "cradle pads" files refer to templates for the pads, to make cutting easier. Print them out and cut around them on the foam, and the pads will be correct dimensions. 
rsdocx files are DesignSpark Mechanical (CAD software) native files.